Major revamp to how unicode and whitespace is handled.
Previously, whitespace was implied based on the semantics of different
tokens/nodes. This was majorly stupid, ended up creating the need for
special whitespace tokens and nodes, as well as fake empty strings
and all sorts of shenanigans. Instead, each token (and node) now just
stores a slice of all the whitespace that proceeded it from the
source-code itself.

We also actually read unicode (utf8) correctly, instead of having the
lexer basically just break on non-ASCII text.